Refactor chat components to improve code consistency and readability; standardize import statements and formatting across multiple files, including chat.tsx, conversation.tsx, message.tsx, and others. Enhance user experience by ensuring proper handling of chat states and messages.
