+++ app/components/chat/chat.tsx
@@ -1,53 +1,53 @@
-"use client"
+'use client';
 
-import { ChatInput } from "@/app/components/chat-input/chat-input"
-import { Conversation } from "@/app/components/chat/conversation"
-import { useModel } from "@/app/components/chat/use-model"
-import { useChatDraft } from "@/app/hooks/use-chat-draft"
-import { useChats } from "@/lib/chat-store/chats/provider"
-import { useMessages } from "@/lib/chat-store/messages/provider"
-import { useChatSession } from "@/lib/chat-store/session/provider"
-import { SYSTEM_PROMPT_DEFAULT } from "@/lib/config"
-import { useUserPreferences } from "@/lib/user-preference-store/provider"
-import { useUser } from "@/lib/user-store/provider"
-import { cn } from "@/lib/utils"
-import { AnimatePresence, motion } from "motion/react"
-import dynamic from "next/dynamic"
-import { redirect } from "next/navigation"
-import { useMemo, useState } from "react"
-import { useChatCore } from "./use-chat-core"
-import { useChatOperations } from "./use-chat-operations"
-import { useFileUpload } from "./use-file-upload"
+import { ChatInput } from '@/app/components/chat-input/chat-input';
+import { Conversation } from '@/app/components/chat/conversation';
+import { useModel } from '@/app/components/chat/use-model';
+import { useChatDraft } from '@/app/hooks/use-chat-draft';
+import { useChats } from '@/lib/chat-store/chats/provider';
+import { useMessages } from '@/lib/chat-store/messages/provider';
+import { useChatSession } from '@/lib/chat-store/session/provider';
+import { SYSTEM_PROMPT_DEFAULT } from '@/lib/config';
+import { useUserPreferences } from '@/lib/user-preference-store/provider';
+import { useUser } from '@/lib/user-store/provider';
+import { cn } from '@/lib/utils';
+import { AnimatePresence, motion } from 'motion/react';
+import dynamic from 'next/dynamic';
+import { redirect } from 'next/navigation';
+import { useMemo, useState } from 'react';
+import { useChatCore } from './use-chat-core';
+import { useChatOperations } from './use-chat-operations';
+import { useFileUpload } from './use-file-upload';
 
 const FeedbackWidget = dynamic(
-  () => import("./feedback-widget").then((mod) => mod.FeedbackWidget),
+  () => import('./feedback-widget').then((mod) => mod.FeedbackWidget),
   { ssr: false }
-)
+);
 
 const DialogAuth = dynamic(
-  () => import("./dialog-auth").then((mod) => mod.DialogAuth),
+  () => import('./dialog-auth').then((mod) => mod.DialogAuth),
   { ssr: false }
-)
+);
 
 export function Chat() {
-  const { chatId } = useChatSession()
+  const { chatId } = useChatSession();
   const {
     createNewChat,
     getChatById,
     updateChatModel,
     bumpChat,
     isLoading: isChatsLoading,
-  } = useChats()
+  } = useChats();
 
   const currentChat = useMemo(
     () => (chatId ? getChatById(chatId) : null),
     [chatId, getChatById]
-  )
+  );
 
-  const { messages: initialMessages, cacheAndAddMessage } = useMessages()
-  const { user } = useUser()
-  const { preferences } = useUserPreferences()
-  const { draftValue, clearDraft } = useChatDraft(chatId)
+  const { messages: initialMessages, cacheAndAddMessage } = useMessages();
+  const { user } = useUser();
+  const { preferences } = useUserPreferences();
+  const { draftValue, clearDraft } = useChatDraft(chatId);
 
   // File upload functionality
   const {
@@ -58,7 +58,7 @@ export function Chat() {
     cleanupOptimisticAttachments,
     handleFileUpload,
     handleFileRemove,
-  } = useFileUpload()
+  } = useFileUpload();
 
   // Model selection
   const { selectedModel, handleModelChange } = useModel({
@@ -66,15 +66,15 @@ export function Chat() {
     user,
     updateChatModel,
     chatId,
-  })
+  });
 
   // State to pass between hooks
-  const [hasDialogAuth, setHasDialogAuth] = useState(false)
-  const isAuthenticated = useMemo(() => !!user?.id, [user?.id])
+  const [hasDialogAuth, setHasDialogAuth] = useState(false);
+  const isAuthenticated = useMemo(() => !!user?.id, [user?.id]);
   const systemPrompt = useMemo(
     () => user?.system_prompt || SYSTEM_PROMPT_DEFAULT,
     [user?.system_prompt]
-  )
+  );
 
   // Chat operations (utils + handlers) - created first
   const { checkLimitsAndNotify, ensureChatExists, handleDelete, handleEdit } =
@@ -88,7 +88,7 @@ export function Chat() {
       setHasDialogAuth,
       setMessages: () => {},
       setInput: () => {},
-    })
+    });
 
   // Core chat functionality (initialization + state + actions)
   const {
@@ -120,7 +120,7 @@ export function Chat() {
     selectedModel,
     clearDraft,
     bumpChat,
-  })
+  });
 
   // Memoize the conversation props to prevent unnecessary rerenders
   const conversationProps = useMemo(
@@ -132,7 +132,7 @@ export function Chat() {
       onReload: handleReload,
     }),
     [messages, status, handleDelete, handleEdit, handleReload]
-  )
+  );
 
   // Memoize the chat input props
   const chatInputProps = useMemo(
@@ -175,7 +175,7 @@ export function Chat() {
       setEnableSearch,
       enableSearch,
     ]
-  )
+  );
 
   // Handle redirect for invalid chatId - only redirect if we're certain the chat doesn't exist
   // and we're not in a transient state during chat creation
@@ -184,19 +184,19 @@ export function Chat() {
     !isChatsLoading &&
     !currentChat &&
     !isSubmitting &&
-    status === "ready" &&
+    status === 'ready' &&
     messages.length === 0 &&
     !hasSentFirstMessageRef.current // Don't redirect if we've already sent a message in this session
   ) {
-    return redirect("/")
+    return redirect('/');
   }
 
-  const showOnboarding = !chatId && messages.length === 0
+  const showOnboarding = !chatId && messages.length === 0;
 
   return (
     <div
       className={cn(
-        "@container/main relative flex h-full flex-col items-center justify-end md:justify-center"
+        '@container/main relative flex h-full flex-col items-center justify-end md:justify-center'
       )}
     >
       <DialogAuth open={hasDialogAuth} setOpen={setHasDialogAuth} />
@@ -228,7 +228,7 @@ export function Chat() {
 
       <motion.div
         className={cn(
-          "relative inset-x-0 bottom-0 z-50 mx-auto w-full max-w-3xl"
+          'relative inset-x-0 bottom-0 z-50 mx-auto w-full max-w-3xl'
         )}
         layout="position"
         layoutId="chat-input-container"
@@ -243,5 +243,5 @@ export function Chat() {
 
       <FeedbackWidget authUserId={user?.id} />
     </div>
-  )
+  );
 }
