+++ app/components/chat/message.tsx
@@ -1,28 +1,28 @@
-import { Message as MessageType } from "@ai-sdk/react"
-import React, { useState } from "react"
-import { MessageAssistant } from "./message-assistant"
-import { MessageUser } from "./message-user"
+import { Message as MessageType } from '@ai-sdk/react';
+import React, { useState } from 'react';
+import { MessageAssistant } from './message-assistant';
+import { MessageUser } from './message-user';
 
 type MessageProps = {
-  variant: MessageType["role"]
-  children: string
-  id: string
-  attachments?: MessageType["experimental_attachments"]
-  isLast?: boolean
-  onDelete: (id: string) => void
-  onEdit: (id: string, newText: string) => void
-  onReload: () => void
-  hasScrollAnchor?: boolean
-  parts?: MessageType["parts"]
-  status?: "streaming" | "ready" | "submitted" | "error"
-  className?: string
-  agent_id?: string | null
+  variant: MessageType['role'];
+  children: string;
+  id: string;
+  attachments?: MessageType['experimental_attachments'];
+  isLast?: boolean;
+  onDelete: (id: string) => void;
+  onEdit: (id: string, newText: string) => void;
+  onReload: () => void;
+  hasScrollAnchor?: boolean;
+  parts?: MessageType['parts'];
+  status?: 'streaming' | 'ready' | 'submitted' | 'error';
+  className?: string;
+  agent_id?: string | null;
   agent?: {
-    id: string
-    name: string
-    avatar_url: string | null
-  } | null
-}
+    id: string;
+    name: string;
+    avatar_url: string | null;
+  } | null;
+};
 
 export function Message({
   variant,
@@ -40,15 +40,15 @@ export function Message({
   agent_id,
   agent,
 }: MessageProps) {
-  const [copied, setCopied] = useState(false)
+  const [copied, setCopied] = useState(false);
 
   const copyToClipboard = () => {
-    navigator.clipboard.writeText(children)
-    setCopied(true)
-    setTimeout(() => setCopied(false), 500)
-  }
+    navigator.clipboard.writeText(children);
+    setCopied(true);
+    setTimeout(() => setCopied(false), 500);
+  };
 
-  if (variant === "user") {
+  if (variant === 'user') {
     return (
       <MessageUser
         copied={copied}
@@ -63,10 +63,10 @@ export function Message({
       >
         {children}
       </MessageUser>
-    )
+    );
   }
 
-  if (variant === "assistant") {
+  if (variant === 'assistant') {
     return (
       <MessageAssistant
         copied={copied}
@@ -82,8 +82,8 @@ export function Message({
       >
         {children}
       </MessageAssistant>
-    )
+    );
   }
 
-  return null
+  return null;
 }
