+++ app/share/[chatId]/header.tsx
@@ -1,9 +1,9 @@
-import Link from "next/link"
-import React from "react"
+import Link from 'next/link';
+import React from 'react';
 
 export function Header() {
   return (
-    <header className="h-app-header fixed top-0 right-0 left-0 z-50">
+    <header className="h-app-header fixed top-0 right-0 left-0 z-[9999]">
       <div className="h-app-header top-app-header bg-background pointer-events-none absolute left-0 z-50 mx-auto w-full to-transparent backdrop-blur-xl [-webkit-mask-image:linear-gradient(to_bottom,black,transparent)] lg:hidden"></div>
       <div className="bg-background relative mx-auto flex h-full max-w-6xl items-center justify-between px-4 sm:px-6 lg:bg-transparent lg:px-8">
         <Link href="/" className="text-xl font-medium tracking-tight">
@@ -11,5 +11,5 @@ export function Header() {
         </Link>
       </div>
     </header>
-  )
+  );
 }
